pyproject.toml
[project]
name = "donation-box-e3dc-plugin"
version = "0.1.0"
description = "Mega tolle SpendenBox"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "argon2-cffi>=23.1.0",
    "e3dc>=0.1.0",
    "fastapi[standard]>=0.115.6",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
    "ruff>=0.7.3",
]

[tool.uv.sources]
e3dc = { workspace = true }

[tool.uv.workspace]
members = ["packages/*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/donationbox"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = ["E501"]
